starter-packs/gitlab-connector-pack/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-connector-pack"
version = "1.0.0"
description = "Interact with GitLab projects, merge requests, and pipelines."
requires-python = ">=3.10"
license = "MIT"
dependencies = ["python-gitlab>=4.0"]

[tool.setuptools.packages.find]
where = ["src"]
